dom/canvas/test/test_workercanvas_neuter.html

<!DOCTYPE HTML>
<html>
<head>
<title>WorkerCanvas: Test neutering</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<canvas id="c" width="64" height="64"></canvas>
<script>

SimpleTest.waitForExplicitFinish();

function runTest() {

  var htmlCanvas = document.getElementById("c");
  var worker = new Worker("workercanvas_neuter.js");

  ok(htmlCanvas, "Should have HTML canvas element");
  ok(worker, "Web worker successfully created");

  ok(htmlCanvas.transferControlToWorker, "HTMLCanvasElement has transferControlToWorker function");

  var workerCanvas = htmlCanvas.transferControlToWorker();
  ok(workerCanvas, "Expected transferControlToWorker to succeed");

  /* check html canvas is neuterd */
  is(htmlCanvas.width, 64, "HTML canvas has correct width");
  SimpleTest.doesThrow(
    function() { htmlCanvas.width = 128; },
    "Can't change html canvas' width after transferControlToWorker");

  SimpleTest.doesThrow(
    function() { htmlCanvas.height = 128; },
    "Can't change html canvas' height after transferControlToWorker");

  ok(!htmlCanvas.getContext("2d"), "Can't getContext after transferControlToWorker");

  worker.postMessage(workerCanvas, [workerCanvas]);

  /* check parent workercanvas is neutered after being transfered */
  SimpleTest.doesThrow(
    function() { workerCanvas.width = 128; },
    "Can't change transfered worker canvas width");

  SimpleTest.doesThrow(
    function() { workerCanvas.height = 128; },
    "Can't change transfered worker canvas height");

  SimpleTest.doesThrow(
    function() { workerCanvas.getContext("webgl") },
    "Can't getContext on transfered worker canvas");

  // Transfer a neutered workercanvas should be ok.
  worker.postMessage(workerCanvas, [workerCanvas]);

  worker.terminate();
  SimpleTest.finish();
}

SpecialPowers.pushPrefEnv({'set': [
  ['gfx.workercanvas.enabled', true],
  ['webgl.force-enabled', true],
]}, runTest);

</script>
</body>
</html>
